+++ webgoat-integration-tests/src/test/resources/application-inttest.properties
@@ -1,6 +1,10 @@
+#In order to run tests a known temp directory is preferred
+#that is why these values are used
+
 webgoat.user.directory=${user.dir}/target/.webgoat
 webgoat.server.directory=${user.dir}/target/.webgoat
-
 webwolf.fileserver.location=${user.dir}/target/webwolf-fileserver
 
-spring.jpa.hibernate.ddl-auto=create-drop
+#database will get deleted for every mvn clean install
+#as these extra properties are read by WebGoat and WebWolf the drop of the tables 
+#was not helpful.
